svg/interact/scripted/async-01.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>async on script</title><metadata><h:link href="https://svgwg.org/svg2-draft/interact.html#ScriptElement" rel="help"/></metadata><h:meta content="long" name="timeout"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script>let t = async_test(&quot;An async script does not block the parser while downloading&quot;); let results = []; t.step_timeout(function() { assert_equals(JSON.stringify(results), &apos;[&quot;2&quot;,&quot;1&quot;]&apos;) t.done(); }, 4000); function log(text) { results.push(text); }</script><script async="true" href="log.py?sec=3&amp;id=1"/><script>log(&apos;2&apos;);</script></svg>
