LayoutTests/imported/w3c/web-platform-tests/svg/interact/scripted/async-01-expected.txt


FAIL An async script does not block the parser while downloading assert_equals: expected "[\"2\",\"1\"]" but got "[\"1\",\"2\"]"
